+++ src/Proceso.cpp
@@ -3,5 +3,5 @@
 //
 #include "Proceso.h"
 
-Proceso::Proceso(std::string etiqueta, int bt, int at, int q, int pr): etiqueta(etiqueta),BT(bt),AT(at),Q(q),Pr(pr),
-        restante(bt),WT(0),CT(0),RT(-1,TAT(0)),finalizado(false){}
+Proceso::Proceso(std::string etiqueta, double bt, double at, int q, int pr): etiqueta(etiqueta),BT(bt),AT(at),Q(q),Pr(pr),
+        restante(bt),WT(0),CT(0),RT(-1),TAT(0),finalizado(false){}
